requirements.txt
# Core Framework
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20

# Data Validation and Settings
pydantic==2.11.10
pydantic-settings==2.11.0
python-dotenv==1.1.1

# Database
motor==3.7.1
pymongo==4.15.2

# OCR and Image Processing
pytesseract==0.3.13
pillow==11.3.0
PyMuPDF==1.26.4

# NLP and Machine Learning
sentence-transformers==5.1.1
transformers==4.57.0
torch==2.8.0
scikit-learn==1.7.2
numpy==2.3.3

# Utilities
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Monitoring and Logging
structlog==24.4.0